% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemcellindex.R
\name{stemcellindex}
\alias{stemcellindex}
\title{Calculate the Intestinal Stem Cell Index}
\usage{
stemcellindex(gextable, organism, geneid)
}
\arguments{
\item{gextable}{Table of gene expression data, with genes in rows and samples in columns.}

\item{organism}{Organism of source data, either "mouse" or "human"}

\item{geneid}{Type of gene ID in row names, either "ensembl" for Ensembl Gene IDs or "genename" for Gene Names}
}
\value{
A data frame containing RSC scores, CBC scores, and Stem Cell Index
}
\description{
This function calculates the Intestinal Stem Cell Index and single-sample
Regenerative Stem Cell (RSC) and Crypt Base Columnar (CBC) Stem Cell scores.
}
